# cmd/service/docs/swagger.yaml
basePath: /api/v1/measurements
host: localhost:8082
info:
  contact: {}
  description: This is the API for the Meteo Airport project.
  title: Meteo Airport API
  version: "1"
paths:
  /interval/{type}/:
    get:
      consumes:
      - application/json
      description: Get measurements for a specified type within a time range.
      operationId: measurement-interval
      parameters:
      - description: Measurement type
        in: path
        name: type
        required: true
        type: string
      - description: Start date in the format -Hh
        in: query
        name: start
        required: true
        type: string
      - description: End date in the format -Hh
        in: query
        name: end
        required: true
        type: string
      - description: Airport code
        in: query
        name: airport
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get measurements in a specific time interval
      tags:
      - measurements
  /mean/:
    get:
      consumes:
      - application/json
      description: Get the average measurement for specified types on a given date
      operationId: avg-measurement-in-a-day
      parameters:
      - description: Date in the format YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      - description: Comma-separated list of measurement types (e.g., temperature,
          humidity, ...)
        in: query
        name: types
        type: string
      - description: Airport code
        in: query
        name: airport
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get average measurement in a day
      tags:
      - measurements
schemes:
- http
swagger: "2.0"